refactor(ui): B5 -- split RightPanel god-file into LoginPanel / AccountPanel / CompactNewsFeed
RightPanel.kt sat at 809 LOC with four top-level Composables sharing a
single 60-line import block. The orchestrator that picks between
auth states + renders the news feed below is conceptually ~30 LOC;
the rest is widget internals that obscure the right-panel layout for
anyone reading it.

Split keeps the same package (`hivens.ui`) so call sites in AppLayout
need no import changes. Files now:

  RightPanel.kt       53 LOC   orchestrator only
  LoginPanel.kt      428 LOC   login + 2FA dialog + cached-creds path
  AccountPanel.kt    122 LOC   logged-in slot + AuthLoadingSlot
  CompactNewsFeed.kt 268 LOC   feed + skeleton + news-item + helpers

AuthLoadingSlot moves from RightPanel into AccountPanel (it shares the
slot's auth-area styling) and changes from `private` to `internal` so
RightPanel can still route AppState.Loading to it.

Pure file move; logic and behaviour unchanged.
